+++ test/BtCancelMessageTest.cc
@@ -39,13 +39,13 @@ public:
   public:
     int32_t index;
     int32_t begin;
-    uint32_t length;
+    int32_t length;
   public:
     MockBtMessageDispatcher2():index(0),
 			       begin(0),
 			       length(0) {}
 
-    virtual void doCancelSendingPieceAction(int32_t index, int32_t begin, uint32_t length) {
+    virtual void doCancelSendingPieceAction(int32_t index, int32_t begin, int32_t length) {
       this->index = index;
       this->begin = begin;
       this->length = length;
@@ -65,10 +65,10 @@ void BtCancelMessageTest::testCreate() {
   PeerMessageUtil::setIntParam(&msg[9], 256);
   PeerMessageUtil::setIntParam(&msg[13], 1024);
   BtCancelMessageHandle pm = BtCancelMessage::create(&msg[4], 13);
-  CPPUNIT_ASSERT_EQUAL((uint8_t)8, pm->getId());
+  CPPUNIT_ASSERT_EQUAL((int8_t)8, pm->getId());
   CPPUNIT_ASSERT_EQUAL(12345, pm->getIndex());
   CPPUNIT_ASSERT_EQUAL(256, pm->getBegin());
-  CPPUNIT_ASSERT_EQUAL((uint32_t)1024, pm->getLength());
+  CPPUNIT_ASSERT_EQUAL(1024, pm->getLength());
 
   // case: payload size is wrong
   try {
