+++ test/BtPieceMessageTest.cc
@@ -48,9 +48,9 @@ public:
   public:
     int32_t index;
     int32_t begin;
-    uint32_t length;
+    int32_t length;
   public:
-    MockBtMessage2(int32_t index, int32_t begin, uint32_t length):index(index), begin(begin), length(length) {}
+    MockBtMessage2(int32_t index, int32_t begin, int32_t length):index(index), begin(begin), length(length) {}
 
   };
 
@@ -60,7 +60,7 @@ public:
   public:
     virtual BtMessageHandle createRejectMessage(int32_t index,
 						int32_t begin,
-						uint32_t length) {
+						int32_t length) {
       MockBtMessage2Handle msg = new MockBtMessage2(index, begin, length);
       return msg;
     }
@@ -111,11 +111,11 @@ void BtPieceMessageTest::testCreate() {
   PeerMessageUtil::setIntParam(&msg[9], 256);
   memcpy(&msg[13], data, sizeof(data));
   BtPieceMessageHandle pm = BtPieceMessage::create(&msg[4], 11);
-  CPPUNIT_ASSERT_EQUAL((uint8_t)7, pm->getId());
+  CPPUNIT_ASSERT_EQUAL((int8_t)7, pm->getId());
   CPPUNIT_ASSERT_EQUAL(12345, pm->getIndex());
   CPPUNIT_ASSERT_EQUAL(256, pm->getBegin());
   CPPUNIT_ASSERT(memcmp(data, pm->getBlock(), sizeof(data)) == 0);
-  CPPUNIT_ASSERT_EQUAL((uint32_t)2, pm->getBlockLength());
+  CPPUNIT_ASSERT_EQUAL(2, pm->getBlockLength());
 
   // case: payload size is wrong
   try {
@@ -174,7 +174,7 @@ void BtPieceMessageTest::testChokingEvent_allowedFastEnabled() {
   MockBtMessage2* rej = (MockBtMessage2*)btMessageDispatcher->messageQueue.front().get();
   CPPUNIT_ASSERT_EQUAL(1, rej->index);
   CPPUNIT_ASSERT_EQUAL(1024, rej->begin);
-  CPPUNIT_ASSERT_EQUAL((uint32_t)16*1024, rej->length);
+  CPPUNIT_ASSERT_EQUAL(16*1024, rej->length);
 }
 
 void BtPieceMessageTest::testChokingEvent_inAmAllowedIndexSet() {
@@ -259,7 +259,7 @@ void BtPieceMessageTest::testCancelSendingPieceEvent_allowedFastEnabled() {
   MockBtMessage2* rej = (MockBtMessage2*)btMessageDispatcher->messageQueue.front().get();
   CPPUNIT_ASSERT_EQUAL(1, rej->index);
   CPPUNIT_ASSERT_EQUAL(1024, rej->begin);
-  CPPUNIT_ASSERT_EQUAL((uint32_t)16*1024, rej->length);
+  CPPUNIT_ASSERT_EQUAL(16*1024, rej->length);
 }
 
 void BtPieceMessageTest::testCancelSendingPieceEvent_invalidate() {
